moree debug

+++ g2o_test.py
@@ -69,9 +69,9 @@ class PoseGraphOptimization(g2o.SparseOptimizer):
                 self.add_vertex(i, t)
 
                 # Add edges between current odom point and all corresponding lidar points
-                start_index = ((f)/2) * 361
+                start_index = int((f/2) * 361)
                 for x in range(361):
-                    lidar_pt = lidar[f/2][x][0]
+                    lidar_pt = lidar[int(f/2)][x][0]
                     # lidar_pt = vertices[int(x+start_index)].estimate()
                     # lidar_pt = self.vertex(super(), x+start_index).estimate()
                     # diff = t - lidar_pt.estimate()
